+++ src/_internals/sdkVersion.js
@@ -1,0 +1,7 @@
+/**
+ * JS file is needed because TypeScript can't use JSON imports
+ * for UMD builds.
+ */
+import { version } from '../../package.json';
+
+export default version;